School-ASP
The project is a REST API developed in .NETCore using the Entity Framework and applying the SOLID programming principles as well as using ORM. The database used in the project was realized using the Code First approach. The project itself was realized as a multi-layered project consisting of five projects, four are class libraries and the fifth is a Web API project. Domain is the first project that is a class library that contains all the entities created in the database via the Code First approach. This project has no connection to other projects and is completely independent of them. EfDataAccess is a class library project and has a link only to the Domain project. On this project, with Nuget Package Manager, packages are installed that allow you to work with the database. Project contains the Configuration and Migration folder, as well as the Context itself, which enables communication with the database. In the Configuration folder, the configurations of all entities are packed, and in the Migration folder, all files that represent the creation of the database and what was subsequently changed in the tables. The Application project is a class library. The project has links to the Domain project. The project includes folders Commands, DataTransferCommand, DataTransferQuery, Email, Exceptions, Queries, Search as well as interfaces IApplication Actor, ICommand, IUseCaseLogger and UseCaseExecutor. The Commands folder defines interfaces for create, update, delete actions for each entity individually. The Query folder defines interfaces for read and readOne actions for each entity individually. The DataTransferCommand folder defines the objects, required by the interfaces from the Command folder and the Query folder, for transferring objects . The Email folder defines the IEmailSender interface, which is necessary for the action of sending e-mail. Exceptions are defined in the Exceptions folder. In the Searches folder, search objects that are necessary for paginations are defined. The Implementation project is a class library. The project has a link to the Application and EfDataAccess projects. On this project, NuGet Package Manager has installed packages AutoMapper and FluentValidation, which facilitates validation. The project includes folders Commands, Email, Queries, Logging, Validators. The Commands folder defines classes that represent the implementations of the interfaces from the Application project using Entity Framework and thus define the executable logic of the interface. The Query folder defines classes that represent the implementation of the interface from the Application project using Entity Framework and thus define the executable logic of the interface. There is a class in the project that represents the implementation of the IEmailSender interface from the Email folder, Application layer. The Logging folder contains the implementation of the IUseCaseLogger interface and is an action that is performed after logging in. The Validators folder contains validation files for each of the entities required for create and read actions. The Api project is an ASP.Net CoreWeb Application of an API. The project is linked to the Domain, EfDataAccess, Application and Implementation projects. On this project, with the help of NuGetPackageManager, JwtBearer package is installed, which serves for authorization using tokens, as well as Swashbuckle. The project includes folders Controllers, Core and wwwroot. The Controllers folder contains controllers of all entities and CRUD actions implemented in them according to already defined methods, as well as RegisterUserController which is used for user registration, TokkenController which is used for logging in and UploadController which is used for file upload. The Core folder contains the profiles required for mapping, the AnonymousActor class is required for registration, the JwtManager and JwtActor classes required for logging in, the GlobalExeceptionHandler for exceptions handling. The wwwroot folder contains the images folder and the images stored in it by implementing the file upload mechanism. In the root of the project there is a Startup file which represents the Dependency Container which delivers to the project all the necessary objects for the realization of all its functionalities.
Technologies – .NETCore, C#, REST API, Entity Framework, DbContext, AutoMapper, Dependency Container, Interface, Swashbuckle, JwtBearer, Code First, Fluent Validation, SQL